Several new features

+++ src/main/java/org/craftercms/engine/service/context/SiteContextRegistry.java
@@ -17,20 +17,20 @@
 package org.craftercms.engine.service.context;
 
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import javax.annotation.PreDestroy;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.BeanPostProcessor;
 
 /**
  * Registry for site {@link SiteContext}s.
  *
  * @author Alfonso Vásquez
  */
-public class SiteContextRegistry implements BeanPostProcessor {
+public class SiteContextRegistry {
 
     private static final Log logger = LogFactory.getLog(SiteContextRegistry.class);
 
@@ -40,18 +40,14 @@ public class SiteContextRegistry implements BeanPostProcessor {
         registry = new ConcurrentHashMap<>();
     }
 
-    @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
-        return bean;
-    }
-
-    @Override
-    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof SiteContext) {
-            register((SiteContext) bean);
+    @PreDestroy
+    public void destroy() {
+        for (Iterator<SiteContext> iter = registry.values().iterator(); iter.hasNext();) {
+            iter.next().destroy();
+            iter.remove();
         }
 
-        return bean;
+        logger.info("Site context registry destroyed");
     }
 
     public Collection<SiteContext> list() {
@@ -65,17 +61,18 @@ public class SiteContextRegistry implements BeanPostProcessor {
     public void register(SiteContext context) {
         registry.put(context.getSiteName(), context);
 
-        logger.info("[SITE CONTEXT REGISTRY] Site context registered: " + context);
+        logger.info("Site context registered: " + context);
     }
 
     public SiteContext unregister(String siteName) {
         SiteContext context = registry.remove(siteName);
+        if (context != null) {
+            context.destroy();
 
-        context.destroy();
+            logger.info("Site context unregistered: " + context);
+        }
 
-        logger.info("[SITE CONTEXT REGISTRY] Site context unregistered: " + context);
-
-        return context;
+        return null;
     }
 
 }
